uint16

+++ packets/general.go
@@ -2,6 +2,6 @@ package packets
 
 // From Client, From Server
 type Packet struct {
-   ID      uint8
-   Content []byte
+	ID      uint16
+	Content []byte
 }
